requirements.txt
streamlit
pandas
plotly-express
openpyxl
odfpy